@@ -59,57 +59,59 @@
   "license": "MIT",
   "dependencies": {
     "animate.css": "3.4.0",
-    "autobahn": "0.9.9",
+    "autobahn": "17.5.2",
     "babel-eslint": "^7.2.3",
     "babel-plugin-transform-class-properties": "^6.24.1",
     "babel-plugin-transform-decorators-legacy": "1.3.4",
     "babel-preset-flow": "^6.23.0",
     "babel-preset-react": "^6.24.1",
-    "browser-redux-sync": "1.1.0",
     "classnames": "2.2.3",
     "d3": "3.5.12",
     "d3-time-format": "2.0.5",
-    "fixed-data-table": "0.6.3",
+    "fixed-data-table": "0.6.5",
     "font-awesome": "4.5.0",
     "lodash": "3.10.1",
     "lodash.keyby": "4.6.0",
     "moment": "2.13.0",
     "numeral": "1.5.3",
     "nvd3": "1.8.5",
-    "react": "15.6.1",
-    "react-dom": "15.6.1",
+    "prop-types": "15.6.0",
+    "react": "16.2.0",
+    "react-dom": "16.2.0",
     "react-nvd3": "0.5.7",
     "react-redux": "5.0.6",
-    "react-sizeme": "2.3.4",
+    "react-sizeme": "2.3.6",
     "react-sparklines": "1.7.0",
     "redux": "3.7.2",
     "redux-actions": "^2.2.1",
-    "redux-observable": "0.14.1",
-    "redux-persist": "4.8.3",
-    "redux-persist-crosstab": "3.5.3",
+    "redux-observable": "0.17.0",
+    "redux-persist": "5.6.5",
+    "redux-persist-crosstab": "3.7.0",
     "rxjs": "5.4.2",
     "ua-parser-js": "0.7.10",
     "ws": "0.8.1"
   },
   "devDependencies": {
-    "@storybook/react": "^3.1.9",
-    "@types/autobahn": "^0.9.35",
-    "@types/jest": "20.0.5",
-    "@types/karma": "0.13.36",
-    "@types/lodash": "3.10.1",
-    "@types/react": "16.0.2",
-    "@types/react-dom": "15.5.2",
-    "@types/react-virtualized": "9.7.3",
-    "@types/redux-actions": "1.2.7",
+    "@storybook/react": "3.3.11",
+    "@types/autobahn": "^0.9.38",
+    "@types/jest": "22.1.1",
+    "@types/karma": "1.7.3",
+    "@types/lodash": "4.14.100",
+    "@types/numeral": "0.0.22",
+    "@types/openfin": "17.0.4",
+    "@types/react": "16.0.36",
+    "@types/react-dom": "16.0.3",
+    "@types/react-virtualized": "9.7.10",
+    "@types/redux-actions": "2.2.3",
     "@types/storybook__addon-actions": "3.0.1",
-    "@types/storybook__addon-links": "3.0.1",
-    "@types/storybook__react": "3.0.2",
-    "autoprefixer": "6.3.1",
-    "awesome-typescript-loader": "3.2.2",
-    "babel-core": "^6.25.0",
-    "babel-jest": "^20.0.3",
-    "babel-loader": "^7.1.1",
-    "babel-plugin-transform-es2015-modules-commonjs": "6.24.1",
+    "@types/storybook__addon-links": "3.0.2",
+    "@types/storybook__react": "3.0.6",
+    "autoprefixer": "7.2.5",
+    "awesome-typescript-loader": "3.4.1",
+    "babel-core": "^6.26.0",
+    "babel-jest": "^22.1.0",
+    "babel-loader": "^7.1.2",
+    "babel-plugin-transform-es2015-modules-commonjs": "6.26.0",
     "babel-preset-env": "^1.6.0",
     "babel-preset-es2015": "6.24.1",
     "babel-preset-stage-2": "6.24.1",
@@ -118,13 +120,15 @@
     "copy-webpack-plugin": "0.3.3",
     "css-loader": "^0.28.1",
     "dotenv": "1.2.0",
-    "enzyme": "2.9.1",
-    "eslint": "^4.3.0",
-    "eslint-config-airbnb": "15.1.0",
+    "enzyme": "^3.3.0",
+    "eslint": "4.16.0",
+    "eslint-config-airbnb": "16.1.0",
+    "eslint-plugin-import": "2.8.0",
+    "eslint-plugin-jsx-a11y": "6.0.3",
     "eslint-loader": "1.9.0",
-    "eslint-plugin-react": "7.1.0",
+    "eslint-plugin-react": "7.6.1",
     "exports-loader": "0.6.3",
-    "extract-text-webpack-plugin": "^2.1.2",
+    "extract-text-webpack-plugin": "^3.0.2",
     "file-loader": "^0.11.2",
     "html-webpack-plugin": "^2.29.0",
     "install": "0.6.1",
@@ -133,9 +137,8 @@
     "json-loader": "0.5.4",
     "node-sass": "^4.5.3",
     "npm-run-all": "2.1.0",
-    "openfin-cli": "1.0.2",
-    "react-addons-test-utils": "15.6.0",
-    "react-test-renderer": "15.6.1",
+    "openfin-cli": "1.1.5",
+    "react-test-renderer": "16.2.0",
     "react-transform-catch-errors": "1.0.2",
     "react-transform-hmr": "1.0.4",
     "react-virtualized": "9.9.0",
