@@ -1,5 +1,6 @@
 import * as React from 'react'
 import OpenFinChrome from './OpenFinChrome'
+import * as PropTypes from 'prop-types'
 
 export interface ChildContextTypes {
   getChildContext?: () => boolean
@@ -9,7 +10,7 @@ export interface ChildContextTypes {
 export default class OpenFinProvider extends React.Component<ChildContextTypes, {}> {
   props
   static childContextTypes = {
-    openFin: React.PropTypes.object,
+    openFin: PropTypes.object
   }
 
   getChildContext() {
@@ -21,14 +22,14 @@ export default class OpenFinProvider extends React.Component<ChildContextTypes, 
   render() {
     const { openFin } = this.props
 
-    return openFin.isRunningInOpenFin ? 
+    return openFin.isRunningInOpenFin ?
       <OpenFinChrome
         openFin={openFin}
         minimize={openFin.minimize.bind(openFin)}
         maximize={openFin.maximize.bind(openFin)}
         close={openFin.close.bind(openFin)}>
         {this.props.children}
-      </OpenFinChrome> 
+      </OpenFinChrome>
     : React.Children.only(this.props.children)
   }
 }
