@@ -19,7 +19,7 @@ describe.only('TradeNotification', () => {
       tradeId: 'ABC',
     }
     const component = <TradeNotification
-          className="spot-tile__trade-summary"
+          className='spot-tile__trade-summary'
           notification={notification}
           onDismissedClicked={mockOnDismissedClicked}/>
 
@@ -48,7 +48,7 @@ describe.only('TradeNotification', () => {
         tradeId: 'ABC',
       }
       const component = <TradeNotification
-            className="spot-tile__trade-summary"
+            className='spot-tile__trade-summary'
             notification={notification}
             onDismissedClicked={() => {}}/>
 
